requirements.txt
Django==5.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow==10.1.0
requests==2.31.0
python-decouple==3.8
django-crispy-forms==2.1
crispy-bootstrap4==2022.1
django-extensions==3.2.3
python-slugify==8.0.1
bleach==6.1.0
markdown==3.5.1
gunicorn==21.2.0
whitenoise==6.6.0
psycopg2-binary==2.9.9 